data_pipeline/extraction/news_content/requirements.txt
numpy~=1.26.2
requests~=2.31.0
beautifulsoup4~=4.12.2
nltk~=3.8.1
lxml~=4.9.3
newspaper3k
chardet
